Fixed bug where collector message deleted can crash bot

+++ src/commands/config/ConfigureAfkCheck.ts
@@ -967,6 +967,9 @@ export class ConfigureAfkCheck extends BaseCommand {
      * @param {Message} botMsg The bot message.
      */
     public async dispose(ctx: ICommandContext, botMsg: Message | null): Promise<void> {
+        if (botMsg && !(await GuildFgrUtilities.hasMessage(botMsg.channel, botMsg.id)))
+            return;
+
         await botMsg?.delete().catch();
     }
 }
