+++ src/commands/config/ConfigureDungeons.ts
@@ -2148,6 +2148,8 @@ export class ConfigureDungeons extends BaseCommand {
      * @param {Message} botMsg The bot message.
      */
     public async dispose(ctx: ICommandContext, botMsg: Message | null): Promise<void> {
+        if (botMsg && !(await GuildFgrUtilities.hasMessage(botMsg.channel, botMsg.id)))
+            return;
         await botMsg?.delete().catch();
     }
 }
