+++ src/commands/config/ConfigureQuotas.ts
@@ -1234,6 +1234,8 @@ export class ConfigureQuotas extends BaseCommand {
      * @param {Message} botMsg The bot message.
      */
     public async dispose(ctx: ICommandContext, botMsg: Message | null): Promise<void> {
+        if (botMsg && !(await GuildFgrUtilities.hasMessage(botMsg.channel, botMsg.id)))
+            return;
         await botMsg?.delete().catch();
     }
 }
