+++ src/commands/config/ConfigureRoles.ts
@@ -594,6 +594,8 @@ export class ConfigureRoles extends BaseCommand implements IConfigCommand {
 
     /** @inheritDoc */
     public async dispose(ctx: ICommandContext, botMsg: Message | null, ...args: any[]): Promise<void> {
+        if (botMsg && !(await GuildFgrUtilities.hasMessage(botMsg.channel, botMsg.id)))
+            return;
         await botMsg?.delete().catch();
     }
