+++ src/commands/config/ConfigureVerification.ts
@@ -1358,6 +1358,8 @@ export class ConfigureVerification extends BaseCommand {
      * @param {Message} botMsg The bot message.
      */
     public async dispose(ctx: ICommandContext, botMsg: Message | null): Promise<void> {
+        if (botMsg?.deleted)
+            return;
         await botMsg?.delete().catch();
     }
 }
